+++ src/plugins.ts
@@ -1,5 +1,6 @@
 export {css as cssLR} from '@codemirror/lang-css';
 export {javascript as javascriptLR} from '@codemirror/lang-javascript';
+export {json as jsonLR} from '@codemirror/lang-json';
 export {css} from '@codemirror/legacy-modes/mode/css';
-export {javascript, json} from '@codemirror/legacy-modes/mode/javascript';
+export {javascript} from '@codemirror/legacy-modes/mode/javascript';
 export {lua} from '@codemirror/legacy-modes/mode/lua';